Fix outstanding Flow issues and move TopLevelEventTypes
