+++ packages/react-native-renderer/src/ReactNativeBridgeEventPlugin.js
@@ -12,6 +12,7 @@ import {
   accumulateTwoPhaseDispatches,
   accumulateDirectDispatches,
 } from 'events/EventPropagators';
+import type {TopLevelTypes} from 'events/TopLevelEventTypes';
 import * as ReactNativeViewConfigRegistry from 'ReactNativeViewConfigRegistry';
 import SyntheticEvent from 'events/SyntheticEvent';
 import invariant from 'fbjs/lib/invariant';
@@ -29,7 +30,7 @@ const ReactNativeBridgeEventPlugin = {
    * @see {EventPluginHub.extractEvents}
    */
   extractEvents: function(
-    topLevelType: string,
+    topLevelType: TopLevelTypes,
     targetInst: Object,
     nativeEvent: AnyNativeEvent,
     nativeEventTarget: Object,
