+++ packages/react-native-renderer/src/ReactNativeEventEmitter.js
@@ -15,6 +15,7 @@ import warning from 'fbjs/lib/warning';
 import {getInstanceFromNode} from './ReactNativeComponentTree';
 
 import type {AnyNativeEvent} from 'events/PluginModuleType';
+import type {TopLevelTypes} from 'events/TopLevelEventTypes';
 
 export {getListener, registrationNameModules as registrationNames};
 
@@ -88,7 +89,7 @@ const removeTouchesAtIndices = function(
  */
 export function _receiveRootNodeIDEvent(
   rootNodeID: number,
-  topLevelType: string,
+  topLevelType: TopLevelTypes,
   nativeEventParam: ?AnyNativeEvent,
 ) {
   const nativeEvent = nativeEventParam || EMPTY_NATIVE_EVENT;
@@ -114,7 +115,7 @@ export function _receiveRootNodeIDEvent(
  */
 export function receiveEvent(
   rootNodeID: number,
-  topLevelType: string,
+  topLevelType: TopLevelTypes,
   nativeEventParam: AnyNativeEvent,
 ) {
   _receiveRootNodeIDEvent(rootNodeID, topLevelType, nativeEventParam);
